+++ build.rs
@@ -1,11 +1,14 @@
 #![feature(let_chains)]
 
+use std::cmp::Ordering;
+use std::collections::VecDeque;
 use std::env;
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
 
 use glob::glob;
+use itertools::Itertools;
 
 // Rust has no built-in parametrize feature. We can't use snapshot crates, because key order is non-deterministic.
 // See https://github.com/la10736/rstest/issues/163
@@ -49,20 +52,30 @@ fn prepare_{name}() {{
     for entry in glob("tests/fixtures/indicators/*.jsonl").expect("Failed to read glob pattern") {
         let path = entry.unwrap();
         let name = path.file_stem().unwrap().to_str().unwrap();
-        let function = name.to_ascii_lowercase().replace('-', "_");
-        let mut parts = name.splitn(3, '-');
-        let indicator = parts.next().unwrap();
+        let function = name.to_ascii_lowercase().replace(['-', '+'], "_");
+        let mut parts = name.split('-').collect::<VecDeque<_>>();
+        let ident = parts.pop_front().unwrap();
 
-        let setting = if let Some(field) = parts.next()
-            && let Some(value) = parts.next()
-        {
-            if value.as_bytes().iter().all(u8::is_ascii_digit) {
-                format!("indicators::{indicator} {{ {field}: Some({value}), ..Default::default() }}")
-            } else {
-                format!("indicators::{indicator} {{ {field}: Some(String::from(\"{value}\")), ..Default::default() }}")
+        let setting = match parts.len().cmp(&2) {
+            Ordering::Less => "Default::default()".into(),
+            Ordering::Equal => {
+                let field = parts[0];
+                let value = parts[1].replace('+', "|");
+                if value.as_bytes().iter().all(u8::is_ascii_digit) {
+                    format!("indicators::{ident} {{ {field}: Some({value}), ..Default::default() }}")
+                } else {
+                    format!("indicators::{ident} {{ {field}: Some(String::from(\"{value}\")), ..Default::default() }}")
+                }
             }
-        } else {
-            "Default::default()".into()
+            Ordering::Greater => {
+                let field = parts.pop_front().unwrap();
+                let tuples = parts
+                    .iter()
+                    .tuples()
+                    .map(|(key, value)| format!("(String::from(\"{key}\"), {value})"))
+                    .join(",");
+                format!("indicators::{ident} {{ {field}: Some(HashMap::from([{tuples}])), ..Default::default() }}")
+            }
         };
 
         write!(
@@ -71,7 +84,7 @@ fn prepare_{name}() {{
 #[test]
 fn {function}() {{
     check_indicators("indicators/{name}", Settings {{
-        {indicator}: Some({setting}),
+        {ident}: Some({setting}),
         price_comparison_procurement_methods: Some(String::from("R024|R028|R036|R058")),
         no_price_comparison_procurement_methods: Some(String::from("EX")),
         ..Default::default()